requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
colorama>=0.4.6
Pillow>=10.0.0
python-chess>=1.999
pytz>=2023.3
lavalink>=5.0.0
yarl>=1.9.0
psutil>=5.9.0
wavelink>=3.0.0
gTTS>=2.3.0
git+https://github.com/darknight156/jishaku
git+https://github.com/Rapptz/discord-ext-menus
aiohttp
discord.py
